+++ tests/tests_unit/test_data/test_map_stream.py
@@ -190,9 +190,21 @@ def test_map_stream_for_list_record(general_data: List[dict]):
 def test_map_stream_data_increase(general_data: List[dict]):
     data = (
         Data(general_data)
-        .filter(lambda record: record.get("batchId") is None)
+        .filter(lambda record: record.get("batchId") is None)  # returns 9
         .map_stream(double_generator)
         .map_stream(event_type_generator)
     )
 
+    for m in data:
+        print(m)
+
     assert len(list(data)) == 18
+
+
+def test_map_stream_data_can_yield_None():
+    def x(s):
+        for m in s:
+            yield None
+
+    d = Data(["a", "b"]).map_stream(x)
+    assert list(d) == [None, None]